Refactor View into View and View.{Login,Initial,Thread,Components}
