+++ Shpadoinklekasten-lib.cabal
@@ -22,6 +22,10 @@ library
                        Types.ViewModel
                        Types.Controller
                        View
+                       View.Login
+                       View.Initial
+                       View.Thread
+                       View.Components
   build-depends:       base ^>=4.12.0.0
                      , aeson
                      , base16-bytestring
